requirements.txt
python-telegram-bot>=22.0
python-dotenv==1.0.0
google-api-python-client==2.84.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
pandas==2.0.0
pytz==2023.3
